backend/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
randomtimestamp==2.2
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
